correct compressor samplerate data type

+++ src/dsp/compressor_core.h
@@ -44,7 +44,7 @@ typedef struct {
 	float detectoravg;
 	float compgain;
 	float maxcompdiffdb;
-	float samplerate;
+	int samplerate;
 	float ang90;
 	float ang90inv;
 } sf_compressor_state_st;
